+++ zcash_client_backend/src/fees/orchard.rs
@@ -3,15 +3,15 @@
 
 use std::convert::Infallible;
 
-use orchard::{BundleProtocol, builder::BundleType};
+use orchard::{builder::BundleType, bundle::BundlePoolRestrictions};
 use zcash_protocol::value::Zatoshis;
 
 pub(crate) fn transactional_action_count(
-    protocol: BundleProtocol,
+    pool_restrictions: BundlePoolRestrictions,
     num_spends: usize,
     num_outputs: usize,
 ) -> Result<usize, &'static str> {
-    BundleType::DEFAULT.num_actions(num_spends, num_outputs, protocol)
+    BundleType::DEFAULT.num_actions(num_spends, num_outputs, pool_restrictions)
 }
 
 /// A trait that provides a minimized view of Orchard-style bundle configuration
@@ -22,8 +22,8 @@ pub trait BundleView<NoteRef> {
     /// The type of inputs of the bundle.
     type Out: OutputView;
 
-    /// Returns the protocol rules for the bundle.
-    fn bundle_protocol(&self) -> BundleProtocol;
+    /// Returns the pool restrictions for the bundle.
+    fn bundle_pool_restrictions(&self) -> BundlePoolRestrictions;
     /// Returns the inputs to the bundle.
     fn inputs(&self) -> &[Self::In];
     /// Returns the outputs of the bundle.
@@ -31,12 +31,12 @@ pub trait BundleView<NoteRef> {
 }
 
 impl<'a, NoteRef, In: InputView<NoteRef>, Out: OutputView> BundleView<NoteRef>
-    for (BundleProtocol, &'a [In], &'a [Out])
+    for (BundlePoolRestrictions, &'a [In], &'a [Out])
 {
     type In = In;
     type Out = Out;
 
-    fn bundle_protocol(&self) -> BundleProtocol {
+    fn bundle_pool_restrictions(&self) -> BundlePoolRestrictions {
         self.0
     }
 
@@ -56,8 +56,8 @@ impl<NoteRef> BundleView<NoteRef> for EmptyBundleView {
     type In = Infallible;
     type Out = Infallible;
 
-    fn bundle_protocol(&self) -> BundleProtocol {
-        BundleProtocol::OrchardPreNu6_3
+    fn bundle_pool_restrictions(&self) -> BundlePoolRestrictions {
+        BundlePoolRestrictions::OrchardNu6_2Only
     }
 
     fn inputs(&self) -> &[Self::In] {